CheckStyle - Вынес Node в отдельный класс
